% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modify_time.R
\docType{data}
\name{data_modify_time}
\alias{data_modify_time}
\title{Data product last modification time}
\format{
A data frame with the following columns:
\itemize{
\item \code{taxa}: The taxa group that the location information can be used for.
Note that some taxa groups may have the same 'plotID' but their
latitude/longitude may differ slightly, which justifies the need of this column.
\item \code{neonDPI}: The NEON data product ID. See \code{neonUtilities:::table_types} for all
available data types and their data product IDs provided by NEON.
\item \code{modify_time}: The last modification time.
}
}
\usage{
data_modify_time
}
\description{
This data frame records the last modify time for each data product in this package.
}
\keyword{datasets}
