% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mosquito.R
\docType{data}
\name{data_mosquito}
\alias{data_mosquito}
\title{Mosquitoes sampled from CO2 traps}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{location_id}: Location id.
\item \code{siteID}: NEON site code.
\item \code{unique_sample_id}: Identity of unique samples (equals \code{sampleID}).
\item \code{subsampleID}: Unique identifier associated with each subsample per sampleID.
\item \code{observation_datetime}: Observation date and time.
\item \code{taxon_id}: Accepted species code, based on one or more sources.
\item \code{taxon_name}: Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{taxon_rank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{variable_name}: The variable name(s) represented by the \code{value} column.
\item \code{value}: Abundance (count per trap hour); \code{NA} for zero-catch traps.
\item \code{unit}: Unit of the values in the \code{value} column ('count per trap hour').
\item \code{nativeStatusCode}: The process by which the taxon became established in the location.
'A': Presumed absent; 'N': Native; 'I': Introduced; 'UNK': Status unknown.
\item \code{proportionIdentified}: Proportion of the total catch that was subsampled and identified.
\item \code{release}: Version of data release by NEON.
\item \code{remarks_sorting}: Technician notes; free text comments accompanying the sorting record.
\item \code{samplingProtocolVersion}: The NEON document number and version where detailed information regarding the sampling method used is available; format 'NEON.DOC.######vX'.
\item \code{sex}: M for male, F for female, U for unknown.
\item \code{sortDate}: Date sample was sorted.
\item \code{trapHours}: Number of hours between trap setting and collecting events.
\item \code{latitude}: The geographic latitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{longitude}: The geographic longitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{elevation}: Elevation (in meters) above sea level.
\item \code{nlcdClass}: National Land Cover Database Vegetation Type Name.
\item \code{plotType}: NEON plot type in which sampling occurred: tower, distributed or gradient.
}
}
\usage{
data_mosquito
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10043.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10043.001}.
}
\details{
To clean the data, we:
\enumerate{
\item Joined \code{mos_trapping} to \code{mos_sorting} to \code{mos_expertTaxonomistIDProcessed}.
\item Filtered to \code{targetTaxaPresent == "Y"}, \code{sampleCondition == "No known compromise"}, and \code{taxonRank != "family"}.
\item Estimated total individuals per subsample = \code{individualCount / proportionIdentified}.
\item Abundance = estimated total individuals / \code{trapHours} (count per trap hour).
}
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_location}). We retained records without a \code{taxon_id} (where \code{value} is \code{NA}) to preserve sampling effort for traps that caught zero mosquitoes.
}
\author{
Natalie Robinson, Daijiang Li
}
\keyword{datasets}
