% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_location.R
\docType{data}
\name{neon_location}
\alias{neon_location}
\title{Information of all locations included in this data package.}
\format{
A data frame with the following columns:
\itemize{
\item \code{location_id}: Location id.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{latitude}: The geographic latitude (in decimal degrees, WGS84) of the
geographic center of the reference area.
\item \code{longitude}: The geographic longitude (in decimal degrees, WGS84) of
the geographic center of the reference area.
\item \code{elevation}: Elevation (in meters) above sea level.
\item \code{nlcdClass}: National Land Cover Database Vegetation Type Name for terrestrial sites.
\item \code{aquaticSiteType}: Type of aquatic systems ('lake', 'river', 'stream').
}
}
\usage{
neon_location
}
\description{
We extracted location information from all
taxanomic groups and saved as one file here.
\emph{Note that some aquatic sites do not have lat/long information though.}
}
\keyword{datasets}
