% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_locations.R
\docType{data}
\name{neon_locations}
\alias{neon_locations}
\title{Information of all locations included in this data package.}
\format{
A data frame with the following columns:
\itemize{
\item \code{domainID}: Unique identifier of the NEON domain.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{nlcdClass}: National Land Cover Database Vegetation Type Name for terrestrial sites.
\item \code{decimalLatitude}: The geographic latitude (in decimal degrees, WGS84) of the
geographic center of the reference area.
\item \code{decimalLongitude}: The geographic longitude (in decimal degrees, WGS84) of
the geographic center of the reference area.
\item \code{geodeticDatum}: Model used to measure horizontal position on the earth.
\item \code{coordinateUncertainty}: The horizontal distance (in meters) from the given
decimalLatitude and decimalLongitude describing the smallest circle containing
the whole of the Location. Zero is not a valid value for this term.
\item \code{elevation}: Elevation (in meters) above sea level.
\item \code{elevationUncertainty}: Uncertainty in elevation values (in meters).
\item \code{taxa}: The taxa group that the location information can be used for.
Note that some taxa groups may have the same 'plotID' but their
latitude/longitude may differ slightly, which justifies the need of this column.
\item \code{neonDPI}: The NEON data product ID. See \code{neonUtilities:::table_types} for all
available data types and their data product IDs provided by NEON.
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{aquaticSiteType}: Type of aquatic systems ('lake', 'river', 'stream').
}
}
\usage{
neon_locations
}
\description{
To reduce the size of this data package, we extracted location information from all
taxanomic groups and saved as one file to avoid duplications.
\emph{Note that some aquatic sites do not have lat/long information though.}
}
\keyword{datasets}
