% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_sites.R
\docType{data}
\name{neon_sites}
\alias{neon_sites}
\title{Site information}
\format{
A data frame with the following columns:
\itemize{
\item \verb{Site Name}: Full site name.
\item \code{siteID}: NEON site code.
\item \verb{Domain Name}: Full domain name.
\item \code{domainID}: Unique identifier of the NEON domain.
\item \code{State}: The state name of the site locates in.
\item \code{Latitude}: Latitude of the site  (in decimal degrees, WGS84).
\item \code{Longitude}: Longitude of the site (in decimal degrees, WGS84).
\item \verb{Site Type}: The type of the site (e.g. Core Terrestrial).
\item \verb{Site Subtype}: Second level site type, for aquatic sites only (e.g. Lake, Wadeable Stream, Non-wadeable River).
\item \verb{Site Host}: Host organization of the site.
}
}
\usage{
neon_sites
}
\description{
Full names, types, coordinates of all 81 NEON sites.
}
\keyword{datasets}
