% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_algae.R
\docType{data}
\name{data_algae}
\alias{data_algae}
\title{Periphyton, seston, and phytoplankton collection}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{siteID}: NEON site code.
\item \code{sampleID}: Identifier for sample.
\item \code{eventID}:	An identifier for the set of information associated with the event,
which includes information about the place and time of the event.
\item \code{namedLocation}:	Name of the measurement location in the NEON database.
\item \code{habitatType}:	Habitat type sampled.
\item \code{taxonID}: Species code, based on one or more sources.
\item \code{scientificName}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{family}: The scientific name of the family in which the taxon is classified.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{collectDate}: Date of the collection event.
\item \code{density}: Density of algae.
\item \code{cell_density_standardized_unit}: Unit of \code{density}.
\item \code{algalParameterValue}: Value of the algalParameter.
\item \code{algalParameter}: Parameter used for analysis of algal assemblages.
\item \code{perBSVol}: Per bottle sample volumn (milliliter).
\item \code{fieldSampleVolume}: Sample volume collected in the field (milliliter).
\item \code{algalSampleType}: Type of algal sample collected.
\item \code{benthicArea}: Area of the benthos sampled (square Meter).
}
}
\usage{
data_algae
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.20166.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.20166.001}.
}
\details{
Here, we:
\itemize{
\item Only records from the ‘alg_biomass’ with the ‘analysis_type’ taxonomy were used for the observation table.
\item Combined the preservative sample volume  and lab’s recorded volumes from the ‘alg_biomass’ file and ‘alg_tax_long’ for total sample volume. If missing ‘perBottleSampleVolume’, they were created using NEON domain lab volumes. This new volume column was labeled ‘perBSVol’.
\item Combined ‘Alg_field_data’ with ‘alg_biomass’ via ‘parentSampleID’ to add in field conditions and ‘benthicArea’.
\item Corrected sample units from ‘cellsperBottle’ to density by dividing ‘algalParameterValue’ updated sample volume ‘perBSVol. Benthic samples units were then corrected to density by multiplying ‘algalParameterValue’ by ‘fieldSampleVolume’ and dividing by ‘benthicArea’ sampled.
}
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\author{
Lara Jansen
}
\keyword{datasets}
