% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_beetle.R
\docType{data}
\name{data_beetle}
\alias{data_beetle}
\title{Ground beetles sampled from pitfall traps}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{sampleID}: Identifier for sample.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{trapID}: Identifier for trap.
\item \code{setDate}: Date that trap was set.
\item \code{collectDate}: Date of the collection event.
\item \code{trappingDays}: Decimal days between trap setting and collecting events.
\item \code{nativeStatusCode}: The process by which the taxon became established in the location.
'A': Presumed absent, due to lack of data indicating a taxon's presence in a given location;
'N': Native; 'I': Introduced; 'UNK': Status unknown.
\item \code{identificationSource}: Source of identification.
\item \code{taxonID}: Species code, based on one or more sources.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{scientificName}: Scientific name, associated with the taxonID. This is the name of the lowest level taxonomic rank that can be determined.
\item \code{count}: Number of individuals. \code{NA} represents no beetles catche in the trap.
}
}
\usage{
data_beetle
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10022.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10022.001}.
}
\details{
BRIEF STEPS WE HAVE TOOK TO PREPARE THIS PRODUCT...
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\keyword{datasets}
