% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_herp_bycatch.R
\docType{data}
\name{data_herp_bycatch}
\alias{data_herp_bycatch}
\title{Vertebrate Herpetofauna Bycatch sampled from pitfall traps}
\format{
A data frame (tibble) with the following columns:
\itemize{
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{trapID}: Identifier for trap.
\item \code{setDate}: Date that trap was set
\item \code{collectDate}: Date of the collection event
\item \code{eventID}: Cleaned up identifier for the set of information associated with the event, which includes information about the place and time of the event
\item \code{trappingDays}: Cleaned up decimal days between trap setting and collecting events
\item \code{taxonID}: Species code, based on one or more sources only provided for vert bycatch herp
\item \code{scientificName}: Scientific name, associated with the taxonID. This is the name of the lowest level taxonomic rank that can be determined only provided for vert bycatch herp
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen only provided for vert bycatch herp
\item \code{individualCount}: Number of individuals of the same type, summed across types for all sampleTypes but vert bycatch herp
\item \code{nativeStatusCode}: 	The process by which the taxon became established in the location only provided for vert bycatch herp
\item \code{remarksSorting}: Technician notes; free text comments accompanying the record from sorting table
\item \code{remarksFielddata}: Technician notes; free text comments accompanying the record from fielddata table
}
}
\source{
\url{https://data.neonscience.org}
}
\usage{
data_herp_bycatch
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10022.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10022.001}.
}
\details{
To process data we:
\enumerate{
\item Cleaned trappingDays.
\itemize{
\item So that is is the number of days a trap was set before being collected.
\item Correct trap days to account for entries where the trap set date was not updated based on a previous collection.
}
\item Create a boutID that identifies all trap collection events at a site in the same bout essentially replacing eventID.
\item Update collectDate to reference the most common collection day in a bout, maintaining one collectDate per bout.
\item sampleType provides the group that was caught in the pit fall trap. This was changed to have three levels
\itemize{
\item "vert bycatch herp" - these are the samples
\item "no data collected" - these samples in fielddata not in sorting
\item "not herp" - this is a aggregate of all the other types "other carabid", "invert bycatch", "carabid", "vert bycatch mam"
And we only kept "vert bycatch herp" in the final data product.
}
}
}
\note{
This script was derived from the script written by Kari Norman to process the pit fall traps of beetles.
Additional variables were added and missing samples were retained in herp_bycatch.
}
\references{
Hoekman, David, Katherine E. LeVan, Cara Gibson, George E. Ball, Robert A. Browne, Robert L. Davidson, Terry L. Erwin, et al. “Design for Ground Beetle Abundance and Diversity Sampling within the National Ecological Observatory Network.” Ecosphere 8, no. 4 (2017): e01744.
}
\author{
Matt Helmus and Kari Norman
}
\keyword{datasets}
