% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_macroinvertebrate.R
\docType{data}
\name{data_macroinvertebrate}
\alias{data_macroinvertebrate}
\title{Macroinvertebrate data}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{siteID}: NEON site code.
\item \code{sampleID}: Identifier for sample.
\item \code{namedLocation}:	Name of the measurement location in the NEON database.
\item \code{collectDate}: Date of the collection event.
\item \code{scientificName}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{acceptedTaxonID}: Accepted species code, based on one or more sources.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{subsamplePercent}: Percent of the total sample contained in the subsample.
\item \code{individualCount}:	Number of individuals of the same type.
\item \code{estimatedTotalCount}: Estimated total count of individuals within a sample,
of given taxon, life stage, and size class.
}
}
\usage{
data_macroinvertebrate
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.20120.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.20120.001}.
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\author{
Stephanie Parker, Eric Sokol
}
\keyword{datasets}
