% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mosquito.R
\docType{data}
\name{data_mosquito}
\alias{data_mosquito}
\title{Mosquitoes sampled from CO2 traps}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{eventID}: An identifier for the set of information associated with the event, which includes information about the place and time of the event.
\item \code{sampleID}: Identifier for sample.
\item \code{taxonID}: Species code, based on one or more sources.
\item \code{scientificName}: Scientific name, associated with the taxonID. This is the name of the lowest level taxonomic rank that can be determined.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{family}: The scientific name of the family in which the taxon is classified.
\item \code{nativeStatusCode}: The process by which the taxon became established in the location.
'A': Presumed absent, due to lack of data indicating a taxon's presence in a given location;
'N': Native; 'I': Introduced; 'UNK': Status unknown.
\item \code{sex}: M for male, F for female, U for unknown.
\item \code{collectDate}: Date of the collection event.
\item \code{startCollectDate}: Earliest known collection date for this sample.
\item \code{endCollectDate}: Latest known collection date for this sample.
\item \code{individualCount}: Number of individuals of the same type.
\item \code{estimated_totIndividuals}: Estimated total number of individuals.
\item \code{trapHours}: Number of hours between trap setting and collecting events.
\item \code{nightOrDay}: Whether sampling occurred primarily during the day or at night.
\item \code{totalWeight}: Weight of entire sample in gram.
\item \code{subsampleWeight}:	Weight of subsample in gram.
}
}
\usage{
data_mosquito
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10043.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10043.001}.
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\keyword{datasets}
