% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mosquito.R
\docType{data}
\name{data_mosquito}
\alias{data_mosquito}
\title{Mosquitoes sampled from CO2 traps}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{location_id}: Location id.
\item \code{siteID}: NEON site code.
\item \code{observation_datetime}: Observation date and time.
\item \code{taxon_id}: Accepted species code, based on one or more sources.
\item \code{taxon_name}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{taxon_rank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{variable_name}: The variable name(s) represented by the \code{value} column.
\item \code{value}: Value of the variable(s) specified by \code{variable_name}.
\item \code{unit}: Unit of the values in the \code{value} column.
\item \code{sortDate}: Date sample was sorted.
\item \code{sampleID}: Identifier for sample.
\item \code{subsampleID}: Unique identifier associated with each subsample per sampleID.
\item \code{totalWeight}: Weight of entire sample in gram.
\item \code{subsampleWeight}:	Weight of subsample in gram.
\item \code{release}: Version of data release by NEON.
\item \code{trapHours}: Number of hours between trap setting and collecting events.
\item \code{samplingProtocolVersion}: The NEON document number and version where detailed information regarding the sampling method used is available; format 'NEON.DOC.######vX'.
\item \code{nativeStatusCode}: The process by which the taxon became established in the location.
'A': Presumed absent, due to lack of data indicating a taxon's presence in a given location;
'N': Native; 'I': Introduced; 'UNK': Status unknown.
\item \code{sex}: M for male, F for female, U for unknown.
\item \code{weightBelowDetection}: Notes regarding the weight relative to scale detection limit.
\item \code{latitude}: The geographic latitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{longitude}: The geographic longitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{elevation}: Elevation (in meters) above sea level.
}
}
\usage{
data_mosquito
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10043.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10043.001}.
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_location}).
}
\author{
Natalie Robinson
}
\keyword{datasets}
