% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plant.R
\docType{data}
\name{data_plant}
\alias{data_plant}
\title{Plant survey data collected by NEON}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{subplotID}: This is the NEON provided subplot ID in the format of \code{subplot_id}, then \code{subsubplot_id} and 1 or 10 (m^2); if the sampling unit is 100 m^2, the values are 31, 32, 40, and 41.
\item \code{boutNumber}: Number of sampling bout, mostly just 1.
\item \code{year}: Observation year.
\item \code{endDate}: The end date-time or interval during which an event occurred.
\item \code{taxonID}: Species code, based on one or more sources.
\item \code{scientificName}: Scientific name, associated with the taxonID. This is the name of the lowest level taxonomic rank that can be determined.
\item \code{taxonRank}: The rank of \code{scientificName}; some of them are genus levels (we filtered out higher ranks already).
\item \code{family}: The family of the species.
\item \code{nativeStatusCode}: Whether the species is a native or non-native species. 'A': Presumed absent, due to lack of data indicating a taxon's presence in a given location; 'N': Native; 'N?': Probably Native; 'I': Introduced; 'I?': Probably Introduced; 'NI': Native and Introduced, some infrataxa are native and others are introduced; 'NI?': Probably Native and Introduced, some infrataxa are native and others are introduced; 'UNK': Status unknown.
\item \code{percentCover}: Ocular estimate of cover of the index (e.g., species) as a percent within 1 m^2 subsubplots.
\item \code{heightPlantOver300cm}: Indicator of whether individuals of the species in the sample are taller than 300 cm.
\item \code{heightPlantSpecies}: Ocular estimate of the height (centimeter) of the plant species.
\item \code{sample_area_m2}: The area of the sampling unit that the observed plant was located in.
\item \code{subplot_id}: Subplot ID; one plot normally has four 100 m^2 subplots (31, 32, 40, 41).
\item \code{subsubplot_id}: Subsubplot ID (1, 2, 3, 4) for sampling units at 1 m^2 or 10 m^2.
}
}
\source{
\url{https://data.neonscience.org}
}
\usage{
data_plant
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10058.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10058.001}
}
\details{
The detailed design of NEON plant survey can be found in \href{https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.2603}{Barnett et al. 2019}.
Here, we:
\enumerate{
\item Removed 1 m^2 data with \code{targetTaxaPresent = N}
\item Removed rows without plotID, subplotID, boutNumber, endDate, and/or taxonID
\item Removed duplicate taxa between nested subplots (each taxon should be represented once for the bout/plotID/year). For example, if a taxon/date/bout/plot combo is present in 1 m^2 data, remove from 10 m^2 and above
\item Stacked species occurrence from different scales into a long data frame. Therefore,
\itemize{
\item to get the species list at 1 m^2 scale, we need all the data with \code{sample_area_m2 == 1} (e.g. \code{dplyr::filter(plants, sample_area_m2 == 1)}); the unique sample unit id can then be generated with \code{paste(plants$plotID, plants$subplot_id, plants$subsubplot_id, sep = "_")}
\item to get the species list at 10 m^2 scale, we need all the data with \code{sample_area_m2 <= 10} (e.g. \code{dplyr::filter(plants, sample_area_m2 <= 10)}); the unique sample unit id can then be generated with \code{paste(plants$plotID, plants$subplot_id, plants$subsubplot_id, sep = "_")}
\item to get the species list at 100 m^2 scale, we need use the whole data set since the maximum value of \code{sample_area_m2} is 100 (i.e. a 10 m by 10 m subplot); the unique sample unit id can then be generated with \code{paste(plants$plotID, plants$subplot_id, sep = "_")}
\item to get the species list at 400 m^2 scale (i.e. one plot with four subplots), we need aggregate the data at \code{plotID} level (the sample unit is the plot now).
}
}
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\references{
Barnett, D.T., Adler, P.B., Chemel, B.R., Duffy, P.A., Enquist, B.J., Grace, J.B., Harrison, S., Peet, R.K., Schimel, D.S., Stohlgren, T.J. and Vellend, M., 2019. The plant diversity sampling design for the national ecological observatory network. Ecosphere, 10(2), p.e02603.
}
\keyword{datasets}
