% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_plant}
\alias{data_plant}
\title{Plant survey data collected by NEON}
\format{
A data frame (also a tibble) with site, plot, latitude, longitude, scientificName, etc.
}
\source{
\url{https://data.neonscience.org}
}
\usage{
data_plant
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10058.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10058.001}
}
\details{
The detailed design of NEON plant survey can be found in \href{https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.2603}{Barnett et al. 2019}.
Here, we:
\enumerate{
\item Removed 1 m^2 data with \code{targetTaxaPresent = N}
\item Removed rows without plotID, subplotID, boutNumber, endDate, and/or taxonID
\item Removed duplicate taxa between nested subplots (each taxon should be represented once for the bout/plotID/year). For example, if a taxon/date/bout/plot combo is present in 1 m^2 data, remove from 10 m^2 and above
\item Stacked species occurrence from different scales into a long data frame. Therefore,
\itemize{
\item to get the species list at 1 m^2 scale, we need all the data with \code{sample_area_m2 == 1} (e.g. \code{dplyr::filter(plants, sample_area_m2 == 1)}); the unique sample unit id can then be generated with \code{paste(plants$plotID, plants$subplot_id, plants$subsubplot_id, sep = "_")}
\item to get the species list at 10 m^2 scale, we need all the data with \code{sample_area_m2 <= 10} (e.g. \code{dplyr::filter(plants, sample_area_m2 <= 10)}); the unique sample unit id can then be generated with \code{paste(plants$plotID, plants$subplot_id, plants$subsubplot_id, sep = "_")}
\item to get the species list at 100 m^2 scale, we need use the whole data set since the maximum value of \code{sample_area_m2} is 100 (i.e. a 10 m by 10 m subplot); the unique sample unit id can then be generated with \code{paste(plants$plotID, plants$subplot_id, sep = "_")}
\item to get the species list at 400 m^2 scale (i.e. one plot with four subplots), we need aggregate the data at \code{plotID} level (the sample unit is the plot now).
}
}
}
\note{
See \code{neonUtilities:::table_types} for all available data types provided by NEON.
}
\references{
Barnett, D.T., Adler, P.B., Chemel, B.R., Duffy, P.A., Enquist, B.J., Grace, J.B., Harrison, S., Peet, R.K., Schimel, D.S., Stohlgren, T.J. and Vellend, M., 2019. The plant diversity sampling design for the national ecological observatory network. Ecosphere, 10(2), p.e02603.
}
\keyword{datasets}
