% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_small_mammal.R
\docType{data}
\name{data_small_mammal}
\alias{data_small_mammal}
\title{Small mammal box trap data collected by NEON}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{trapCoordinate}: Relative coordinate of the trap within the given plotID (A1 - J10). If row or column coordinate is unknown, X is used.
\item \code{trapStatus}: Categorical descriptor of trap status; 0 - no data; 1 - trap not set; 2 - trap disturbed/door closed but empty; 3 - trap door open or closed w/ spoor left; 4 - >1 capture in one trap; 5 - capture; 6 - trap set and empty.
\item \code{year}: Observation year.
\item \code{month}: Observation month.
\item \code{day}: Observation day.
\item \code{taxonID}: Species code, based on one or more sources.
\item \code{scientificName}: Scientific name, associated with the taxonID. This is the name of the lowest level taxonomic rank that can be determined.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{family}: The family of the species.
\item \code{nativeStatusCode}: Whether the species is a native or non-native species.
'A': Presumed absent, due to lack of data indicating a taxon's presence in a
given location; 'N': Native; 'I': Introduced; 'UNK': Status unknown.
\item \code{sex}: M for male, F for female, U for unknown.
\item \code{lifeStage}:	The age class of the individual at the time the Occurrence was recorded.
juvenile = obvious signs of a very young individual, small size, distinctive pelage coloration; subabult; adult.
\item \code{pregnancyStatus}:	Condition at time of capture; if mature: 'pregnant' | 'not'.
\item \code{tailLength}: length of tail; in millimeters.
\item \code{totalLength}: total length (head + body); in millimeters.
\item \code{weight}: Live weight as measured with a spring scale; in grams.
}
}
\source{
\url{https://data.neonscience.org}
}
\usage{
data_small_mammal
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10072.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10072.001}.
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\keyword{datasets}
