% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_small_mammal.R
\docType{data}
\name{data_small_mammal}
\alias{data_small_mammal}
\title{Small mammal box trap data collected by NEON}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{location_id}: Location id.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{unique_sample_id}: Identity of unique samples, usually it has location and date information.
\item \code{observation_datetime}: Observation date and time.
\item \code{taxon_id}: Accepted species code, based on one or more sources.
\item \code{taxon_name}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{taxon_rank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{variable_name}: The variable name(s) represented by the \code{value} column.
\item \code{value}: Value of the variable(s) specified by \code{variable_name}.
\item \code{unit}: Unit of the values in the \code{value} column.
\item \code{year}: Observation year.
\item \code{month}: Observation month.
\item \code{n_trap_nights_per_bout_per_plot}: Number of trap nights per bout per plot.
\item \code{n_nights_per_bout}: Number of nights per bout.
\item \code{nativeStatusCode}: Whether the species is a native or non-native species.
'A': Presumed absent, due to lack of data indicating a taxon's presence in a
given location; 'N': Native; 'I': Introduced; 'UNK': Status unknown.
\item \code{release}: Version of data release by NEON.
\item \code{latitude}: The geographic latitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{longitude}: The geographic longitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{elevation}: Elevation (in meters) above sea level.
\item \code{elevation}: Elevation (in meters) above sea level.
\item \code{plotType}: NEON plot type in which sampling occurred: tower, distributed or gradient.
\item \code{nlcdClass}: National Land Cover Database Vegetation Type Name.
}
}
\source{
\url{https://data.neonscience.org}
}
\usage{
data_small_mammal
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10072.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10072.001}.
}
\details{
To process data we:
\enumerate{
\item Remove all records that are designated as "dead", "escaped", or "nontarget".
\item Remove all records designated as recaptures (i.e., only first captures are retained)
}
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_location}).
}
\author{
Marta Jarzyna
}
\keyword{datasets}
