% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\docType{data}
\name{data_summary}
\alias{data_summary}
\title{Data product last modification time}
\format{
A data frame with the following columns:
\itemize{
\item \code{taxon_group}: The taxa group that the location information can be used for.
Note that some taxa groups may have the same 'plotID' but their
latitude/longitude may differ slightly, which justifies the need of this column.
\item \code{data_package_id}: Identifier for the data generated by \code{{ecocomDP}}.
\item \code{n_taxa}: Number of species included. Note that we did not clean species anmes though
we did remove records with species identified above genus level for well studied groups such as
plant and fish. However, we keep all records regardless of their species identification levels
for taxonomic groups that are hard to identifiy (e.g. macroinvertebrate, beetle).
\item \code{n_sites}: Number of NEON sites included.
\item \code{sites}: All site codes that have data, separated by \code{|}.
\item \code{start_date}: The earliest date that have records.
\item \code{end_date}: The latest date that have records.
\item \code{data_package_title}: Title of the dataset.
\item \code{neon_ecocomdp_mapping_method}: This is the ID that \code{ecocomDP::read_data} used to download and process the NEON data product with NEON's orignal ID specified in \code{original_neon_data_product_id}.
\item \code{original_neon_data_product_id}: The NEON data product ID. See \code{neonUtilities:::table_types} for all
available data types and their data product IDs provided by NEON.
\item \code{original_neon_data_version}: NEON data release version.
\item \code{original_neon_data_doi}: Original NEON data doi.
\item \code{r_object}: The name of R objects for each taxonomic group. By calling the R objects, we can get the processed and standardized NEON organismal data for downstream diversity analysis.
\item \code{variable_names}: The variable names that represent "abundance" information.
\item \code{units}: The units of the \code{variable_names}.
}
}
\usage{
data_summary
}
\description{
This data frame records the last modify time for each data product in this package.
}
\keyword{datasets}
