% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tick.R
\docType{data}
\name{data_tick}
\alias{data_tick}
\title{Ticks sampled using drag cloths}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{collectDate}: Date of the collection event.
\item \code{eventID}: An identifier for the set of information associated with the event, which includes information about the place and time of the event.
\item \code{sampleID}: Identifier for sample.
\item \code{sampleCode}: Barcode of a sample.
\item \code{samplingMethod}: Name or code for the method used to collect or test a sample.
\item \code{totalSampledArea}: Total area sampled (square Meter).
\item \code{targetTaxaPresent}: Indicator of whether the sample contained individuals of the target taxa ('Y' or 'N').
\item \code{remarks_field}: Technician notes; free text comments accompanying the record.
\item \code{acceptedTaxonID}: Accepted species code, based on one or more sources.
\item \code{LifeStage}: Life stage of the sample ('Adult', 'Larva', or 'Nymph').
\item \code{IndividualCount}: Number of individuals of the same type.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{scientificName}: Scientific name, associated with the taxonID. This is the name of the lowest level taxonomic rank that can be determined.
}
}
\usage{
data_tick
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10093.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10093.001}.
}
\details{
BRIEF STEPS WE HAVE TOOK TO PREPARE THIS PRODUCT...
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\keyword{datasets}
