% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tick_pathogen.R
\docType{data}
\name{data_tick_pathogen}
\alias{data_tick_pathogen}
\title{Tick-borne pathogen status}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{namedLocation}: Name of the measurement location in the NEON database.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{collectDate}:	Date of the collection event.
\item \code{endDate}: The end date-time or interval during which an event occurred.
\item \code{hostSpecies}: Host tick species.
\item \code{lifeStage}: Life stage of the host (all Nymph).
\item \code{testedDate}: Date test was conducted.
\item \code{testingID}: Identifier for the group of specimens for testing.
\item \code{batchID}: Identifier for batch or analytical run.
\item \code{testResult}: Result of the test.
\item \code{testPathogenName}	The name of the pathogen.
}
}
\usage{
data_tick_pathogen
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10092.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10092.001}.
}
\details{
To clean the data, we:
\enumerate{
\item Removed samples with flagged quality checks
\item Removed samples with missing metadata (domain/site/plotID; Lat/Long; plotType; nlcdClass, elevation, collectDate, subsampleID, batchID, testingID, testPathogenName
\item Removed samples where sampleCondition!=“OK”
\item Removed samples where testResult==NA
\item Removed samples where HardTick DNA Quality (under testPathogenName) are not ‘Positive’, and then removed hardtack DNA and Ixodes pacificus tests.
\item Combined B. burgdeferi and B burgdeferi sensu lato into a single test pathogen type (B. burgedferi sensu lato)
}
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_locations}).
}
\author{
Melissa Chen, Wynne Moss, Brendan Hobart, Matt Bitters
}
\keyword{datasets}
