% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tick_pathogen.R
\docType{data}
\name{data_tick_pathogen}
\alias{data_tick_pathogen}
\title{Tick-borne pathogen status}
\format{
A data frame (also a tibble) with the following columns:
\itemize{
\item \code{location_id}: Location id.
\item \code{siteID}: NEON site code.
\item \code{plotID}: Plot identifier (NEON site code_XXX).
\item \code{unique_sample_id}: Identity of unique samples, usually it has location and date information.
\item \code{observation_datetime}: Observation date and time.
\item \code{taxon_id}: Accepted species code, based on one or more sources.
\item \code{taxon_name}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{taxon_rank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{variable_name}: The variable name(s) represented by the \code{value} column.
\item \code{value}: Value of the variable(s) specified by \code{variable_name}.
\item \code{unit}: Unit of the values in the \code{value} column.
\item \code{lifeStage}: Life stage of the host (all Nymph).
\item \code{testProtocolVersion}: The protocol version used to test the sample.
\item \code{release}: Version of data release by NEON.
\item \code{n_tests}: Number of tests conducted.
\item \code{n_positive_test}: Number of tests that were positive.
\item \code{latitude}: The geographic latitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{longitude}: The geographic longitude (in decimal degrees, WGS84) of the geographic center of the reference area.
\item \code{elevation}: Elevation (in meters) above sea level.
\item \code{nlcdClass}: National Land Cover Database Vegetation Type Name.
\item \code{plotType}: NEON plot type in which sampling occurred: tower, distributed or gradient.
}
}
\usage{
data_tick_pathogen
}
\description{
This dataset was derived from \href{https://data.neonscience.org}{NEON data portal} with data product ID 'DP1.10092.001'. Details about this data product can be found at \url{https://data.neonscience.org/data-products/DP1.10092.001}.
}
\details{
To clean the data, we:
\enumerate{
\item Removed samples with flagged quality checks
\item Removed samples with missing metadata (domain/site/plotID; Lat/Long; plotType; nlcdClass, elevation, collectDate, subsampleID, batchID, testingID, testPathogenName
\item Removed samples where sampleCondition!=“OK”
\item Removed samples where testResult==NA
\item Removed samples where HardTick DNA Quality (under testPathogenName) are not ‘Positive’, and then removed hardtack DNA and Ixodes pacificus tests.
\item Combined B. burgdeferi and B burgdeferi sensu lato into a single test pathogen type (B. burgedferi sensu lato)
}
}
\note{
Details of locations (e.g. latitude/longitude coordinates can be found in \link{neon_location}).
}
\author{
Melissa Chen, Wynne Moss, Brendan Hobart, Matt Bitters
}
\keyword{datasets}
