% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_taxa.R
\docType{data}
\name{neon_taxa}
\alias{neon_taxa}
\title{Taxanomic names of all groups}
\format{
A data frame with the following columns:
\itemize{
\item \code{taxonID}: Species code, based on one or more sources.
\item \code{acceptedTaxonID}: Accepted species code, based on one or more sources.
\item \code{scientificName}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{taxonRank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{family}: The scientific name of the family in which the taxon is classified.
Not every taxonomic group has this information here.
\item \code{identificationReferences}: A list of sources (concatenated and semicolon separated)
used to derive the specific taxon concept; including field guide editions, books,
or versions of NEON keys used.
\item \code{taxa}: The taxa group that the location information can be used for.
Note that some taxa groups may have the same 'plotID' but their
latitude/longitude may differ slightly, which justifies the need of this column.
\item \code{neonDPI}: The NEON data product ID. See \code{neonUtilities:::table_types} for all
available data types and their data product IDs provided by NEON.
}
}
\usage{
neon_taxa
}
\description{
This data frame was put together from each data product, with the extra column of \code{identificationReferences}.
}
\note{
Some taxonomic groups used \code{taxonID} in the data product while other groups used \code{acceptedTaxonID}.
In addition, all data were from NEON and we did not do extra clean up. For example, sometimes, a species
does not have family information (\code{NA}) while another co-genera species does.
}
\keyword{datasets}
