% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_taxa.R
\docType{data}
\name{neon_taxa}
\alias{neon_taxa}
\title{Taxanomic names of all groups}
\format{
A data frame with the following columns:
\itemize{
\item \code{taxon_id}: Species code, based on one or more sources. For algae, macroinvertebrate, and tick, this is from the \code{acceptedTaxonID} column (which was removed here) so that all taxonomic groups have the same variable name. In another word, algae, macroinvertebrate, and tick only have \code{acceptedTaxonID} and we just renamed it to \code{taxon_id} for these groups following other groups.
\item \code{taxon_name}:	Scientific name, associated with the taxonID. This is the name
of the lowest level taxonomic rank that can be determined.
\item \code{taxon_rank}: The lowest level taxonomic rank that can be determined for the individual or specimen.
\item \code{taxon_group}: The taxonomic group that the location information can be used for.
Note that some taxa groups may have the same 'plotID' but their
latitude/longitude may differ slightly, which justifies the need of this column.
}
}
\usage{
neon_taxa
}
\description{
This data frame was put together from each data product.
}
\note{
Some taxonomic groups used \code{taxonID} (renamed as \code{taxon_id} here) in the data product while other groups used \code{acceptedTaxonID}. In addition, all data were from NEON and we did not do extra clean up.
}
\keyword{datasets}
